Add a "catch-all" mouse button.
This is necessary because the buttons defined by SDL (left, middle, right, x1,
x2) do not include every possible mouse button.

I know this can happen because I pressed a special button on my mouse and it
crashed SDL.Events because there is no matching clause for the extra button.