allow low-level access to buffer in audio callback
The problem with the current type of the audio callback `(CInt -> IO (Vector Word8))`
is that it enforces to allocate a new chunk of memory in every call of the
callback. Audio synthesis is very performance sensitive and every unnecessary
allocation can cause stuttering and hang ups of the audio signal, which is
unacceptable.

I propose to use the type of the low-level SDL Audio API `(Ptr Word8 -> CInt -> IO ())`
which omits allocations. For people who liked the old API, we could offer a
conversion function that converts the high-level callback type to the low-level:
`(CInt -> IO (Vector Word8)) -> (Ptr Word8 -> CInt -> IO ())`