Fix event polling (#242)
pollEvents was apparently broken by a recent-ish
[patch](https://github.com/libsdl-org/SDL/pull/4794) to SDL.

This should fix that. As an aside, this attempts to make polling events
a bit more efficient; We just statically allocate a single buffer for
events (AFAIK you can only call Poll from the main thread anyway, so the
fact that this isn't thread safe shouldn't be an issue).